+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Dropping `#.hasGeneratedKey`.
 - Dropping the `generated` last argument to edge & adjacency iterations.
 - `#.exportEdge` will now always return a key.
+- Fixing mutability bug with `#.copy`.
 - Passing graph instance as second argument to `edgeKeyGenerator`.
 - Improving `#.copy` peformance.